+++ app/Models/Wastetype.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace App\Models;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class Wastetype extends Model
+{
+    use HasFactory;
+
+    /**
+     * fillable
+     * 
+     * @var array
+     */
+    protected $fillable = [
+        'type',
+        'price',
+    ];
+}