+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Api\WastetypeController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -38,4 +39,16 @@ Route::post('/logout', App\Http\Controllers\Api\LogoutController::class)->name('
  */
 Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
+});
+
+/**
+ * route
+ * @method "GET"
+ */
+Route::middleware('auth:api')->group(function () {
+    Route::get('/wastetypes', [WastetypeController::class, 'index'])->name('wastetypes.index');
+    Route::post('/wastetypes', [WastetypeController::class, 'store'])->name('wastetypes.store');
+    Route::get('/wastetypes/{id}', [WastetypeController::class, 'show'])->name('wastetypes.show');
+    Route::put('/wastetypes/{id}', [WastetypeController::class, 'update'])->name('wastetypes.update');
+    Route::delete('/wastetypes/{id}', [WastetypeController::class, 'destroy'])->name('wastetypes.destroy');
 });